refactor(frontend): move Custom Token mapper to lib utils
