+++ src/frontend/src/tests/mocks/ic-tokens.mock.ts
@@ -2,6 +2,9 @@ import { IC_CKBTC_LEDGER_CANISTER_ID } from '$env/networks/networks.icrc.env';
 import type { IcCanisters, IcCkToken, IcToken } from '$icp/types/ic-token';
 import { mockValidToken } from '$tests/mocks/tokens.mock';
 
+export const mockLedgerCanisterId = IC_CKBTC_LEDGER_CANISTER_ID;
+export const mockIndexCanisterId = IC_CKBTC_LEDGER_CANISTER_ID;
+
 export const mockValidIcCanisters: IcCanisters = {
 	ledgerCanisterId: IC_CKBTC_LEDGER_CANISTER_ID
 };
